+++ karaoke-nnice-api/src/main/java/com/nnice/karaoke/repository/CaLamViecRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CaLamViecRepository extends JpaRepository<CaLamViec, Long> {
+public interface CaLamViecRepository extends JpaRepository<CaLamViec, Integer> {
 }
