+++ karaoke-nnice-api/src/main/java/com/nnice/karaoke/repository/KhachHangRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KhachHangRepository extends JpaRepository<KhachHang, Long> {
+public interface KhachHangRepository extends JpaRepository<KhachHang, Integer> {
 }
