+++ karaoke-nnice-api/src/main/java/com/nnice/karaoke/repository/MatHangRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MatHangRepository extends JpaRepository<MatHang, Long> {
+public interface MatHangRepository extends JpaRepository<MatHang, Integer> {
 }
