+++ karaoke-nnice-api/src/main/java/com/nnice/karaoke/repository/PhieuNhapRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PhieuNhapRepository extends JpaRepository<PhieuNhap, Long> {
+public interface PhieuNhapRepository extends JpaRepository<PhieuNhap, Integer> {
 }
