+++ karaoke-nnice-api/src/main/java/com/nnice/karaoke/repository/PhieuSuDungRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PhieuSuDungRepository extends JpaRepository<PhieuSuDung, Long> {
+public interface PhieuSuDungRepository extends JpaRepository<PhieuSuDung, Integer> {
 }
