+++ karaoke-nnice-api/src/main/java/com/nnice/karaoke/repository/PhongRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PhongRepository extends JpaRepository<Phong, Long> {
+public interface PhongRepository extends JpaRepository<Phong, Integer> {
 }
